Rename ProjectGeneratorTask -> TemplateProjectGeneratorTask
This provides a common base task for most of the project generators
