+++ buildSrc/src/main/groovy/org/gradle/testing/performance/generator/tasks/NativeProjectWithDepsGeneratorTask.groovy
@@ -28,7 +28,7 @@ import org.gradle.nativeplatform.NativeLibrarySpec
 /**
  * Generates a multi-project native build that has project dependencies and tests.
  */
-class NativeProjectWithDepsGeneratorTask extends ProjectGeneratorTask {
+class NativeProjectWithDepsGeneratorTask extends TemplateProjectGeneratorTask {
 
     /** Represents a native library requirement */
     class Dependency {
